Add Directory.Build props and targets
When a debug build is running, two properties are added to the project using the props file.
One is to generate a package on build
Second is to include debugging symbols.

After the build, if the build configuration was set to Debug, the nupkg files are copied to a local directory.
The nuget files are references by docker containers as well as projects from the local folder.